requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
pyarrow>=14.0.0
python-dateutil>=2.8.2
tabulate>=0.9.0
pytest>=7.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
openai>=1.15.0
backoff>=2.2.1 